Rename utils to hooks and added hooks folder
