+++ frontend/src/components/sudoku/hooks/useSudokuGrid.js
@@ -11,13 +11,3 @@ export function useSudokuGrid(size) {
 
   return { sudokuGrid, setSudokuGrid, handleCellChange };
 }
-
-export function useSelectedCell() {
-  const [selectedCell, setSelectedCell] = useState({ row: 0, col: 0 });
-
-  const handleCellClick = (row, col) => {
-    setSelectedCell({ row, col });
-  };
-
-  return { selectedCell, setSelectedCell, handleCellClick };
-}
